lltk2/imports.py
# Imports

from django.shortcuts import render, get_object_or_404
from django.core.management.base import BaseCommand
from django.views.generic import View
from django.http import HttpResponse
from django.test import TestCase
from django.contrib import admin
from django.db import models

from django_micro import command, route, template, run, get_app_label