requirements.txt
django-micro
django-extensions
ipython
jupyter